Created: Token service layer
